🚚 More renaming and added URL flag
